With the current redaction, coordinates must be Cartesian.
We could actually use the very same formulae in fractional coordinates,
only loosing the geometrical interpretation of V as the volume of the
tetrahedron. Worth mentioning? It could be more difficult to fix the
weight in fractional coordinates because V would be even more
arbitrary. Need to thing about it...